# Use hdfs.Readdir() to optimize HDFS directory listings