--- a/cmd/gateway/hdfs/gateway-hdfs.go
+++ b/cmd/gateway/hdfs/gateway-hdfs.go
@@ -384,30 +384,12 @@ func (n *hdfsObjects) listDirFactory() minio.ListDirFunc {
 
 // ListObjects lists all blobs in HDFS bucket filtered by prefix.
 func (n *hdfsObjects) ListObjects(ctx context.Context, bucket, prefix, marker, delimiter string, maxKeys int) (loi minio.ListObjectsInfo, err error) {
-	directoryPath := n.hdfsPathJoin(bucket, prefix) + minio.SlashSeparator
-	dirReader, err := n.clnt.Open(directoryPath)
-
-	if err != nil {
-		return loi, hdfsToObjectErr(ctx, err, bucket)
-	}
-
 	fileInfos := make(map[string]os.FileInfo)
-	fileInfos[directoryPath] = dirReader.Stat()
 
-	infos, err := dirReader.Readdir(0)
-
-	if err != nil {
+	if stat, err := n.clnt.Stat(n.hdfsPathJoin(bucket)); err != nil {
 		return loi, hdfsToObjectErr(ctx, err, bucket)
-	}
-
-	for _, fileInfo := range infos {
-		filePath := n.hdfsPathJoin(directoryPath, fileInfo.Name())
-
-		if fileInfo.IsDir() {
-			filePath += minio.SlashSeparator
-		}
-
-		fileInfos[filePath] = fileInfo
+	} else {
+		fileInfos[n.hdfsPathJoin(bucket, prefix) + minio.SlashSeparator] = stat
 	}
 
 	getObjectInfo := func(ctx context.Context, bucket, entry string) (minio.ObjectInfo, error) {
@@ -415,18 +397,50 @@ func (n *hdfsObjects) ListObjects(ctx context.Context, bucket, prefix, marker, d
 		fi, ok := fileInfos[filePath]
 
 		if !ok {
-			err := fmt.Errorf("could not get FileInfo for path '%s'", filePath)
-			return minio.ObjectInfo{}, hdfsToObjectErr(ctx, err, bucket, entry)
+			parentPath := path.Dir(filePath)
+			dirReader, err := n.clnt.Open(parentPath)
+
+			if err != nil {
+				return minio.ObjectInfo{}, hdfsToObjectErr(ctx, err, bucket)
+			}
+
+			fileInfos[parentPath + minio.SlashSeparator] = dirReader.Stat()
+			infos, err := dirReader.Readdir(0)
+
+			if err != nil {
+				return minio.ObjectInfo{}, hdfsToObjectErr(ctx, err, bucket)
+			}
+
+			for _, fileInfo := range infos {
+				filePath := n.hdfsPathJoin(parentPath, fileInfo.Name())
+
+				if fileInfo.IsDir() {
+					filePath += minio.SlashSeparator
+				}
+
+				fileInfos[filePath] = fileInfo
+			}
+
+			fi, ok = fileInfos[filePath]
+
+			if !ok {
+				err = fmt.Errorf("could not get FileInfo for path '%s'", filePath)
+				return minio.ObjectInfo{}, hdfsToObjectErr(ctx, err, bucket, entry)
+			}
 		}
 
-		return minio.ObjectInfo{
+		objectInfo := minio.ObjectInfo{
 			Bucket:  bucket,
 			Name:    entry,
 			ModTime: fi.ModTime(),
 			Size:    fi.Size(),
 			IsDir:   fi.IsDir(),
 			AccTime: fi.(*hdfs.FileInfo).AccessTime(),
-		}, nil
+		}
+
+		delete(fileInfos, filePath)
+
+		return objectInfo, nil
 	}
 
 	return minio.ListObjects(ctx, n, bucket, prefix, marker, delimiter, maxKeys, n.listPool, n.listDirFactory(), getObjectInfo, getObjectInfo)