@blaenk 2020-07-23
It seems like I need to iron out some wrinkles with `--recursive`. Looking into it.