@blaenk 2020-07-24
It's working now with `--recursive` listings.