@blaenk 2020-07-24
Added to OP:

> I also received a report that on a directory of more than 20,000 entries, it reduced ls times from 4 minutes down to 1.27 seconds.