@blaenk 2020-07-24
Happy to help! Thank you both for working on minio.